renaming apply to deploy
